[pp-plans] commented out driving-related plans as they don't handle object poses correctly:
they don't consider that if the robot drives somewhere the perceived object pose in base frame is not valid anymore.
Tricks with TF and going back in time to transform in map and then going back to current time to transform back into base frame seemed to hacky.
Instead, keep the pose in map after perceived and especially before driving away.